# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Social Media Automation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PostStatus = typeof PostStatus[keyof typeof PostStatus];


export const PostStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
  failed: 'failed',
} as const;

export interface Post {
  id: number;
  content: string;
  /** @nullable */
  mediaUrl?: string | null;
  platforms: string[];
  status: PostStatus;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
  /** @nullable */
  errorMessage?: string | null;
}

export interface PostInput {
  /** @minLength 1 */
  content: string;
  mediaUrl?: string;
  /** @minItems 1 */
  platforms: string[];
  scheduledAt?: string;
}

export type PostUpdateStatus = typeof PostUpdateStatus[keyof typeof PostUpdateStatus];


export const PostUpdateStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
} as const;

export interface PostUpdate {
  /** @minLength 1 */
  content?: string;
  mediaUrl?: string;
  platforms?: string[];
  status?: PostUpdateStatus;
  scheduledAt?: string;
}

export type PlatformName = typeof PlatformName[keyof typeof PlatformName];


export const PlatformName = {
  facebook: 'facebook',
  linkedin: 'linkedin',
  twitter: 'twitter',
  instagram: 'instagram',
} as const;

export interface Platform {
  name: PlatformName;
  displayName: string;
  connected: boolean;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  connectedAt?: string | null;
}

export interface PlatformConnectInput {
  accessToken: string;
  accountName?: string;
}

export interface Analytics {
  id: number;
  postId: number;
  platform: string;
  likes: number;
  comments: number;
  shares: number;
  reach: number;
  impressions: number;
  clicks?: number;
  recordedAt: string;
}

export interface AnalyticsSummary {
  totalPosts: number;
  totalPublished: number;
  totalScheduled: number;
  totalDrafts: number;
  totalLikes: number;
  totalComments: number;
  totalShares: number;
  totalReach: number;
  totalImpressions: number;
  postsThisWeek?: number;
  engagementRate?: number;
}

export interface PlatformEngagement {
  platform: string;
  likes: number;
  comments: number;
  shares: number;
  reach: number;
  posts: number;
}

export type ListPostsParams = {
status?: ListPostsStatus;
platform?: string;
};

export type ListPostsStatus = typeof ListPostsStatus[keyof typeof ListPostsStatus];


export const ListPostsStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  published: 'published',
  failed: 'failed',
} as const;

export type GetCalendarPostsParams = {
year: number;
month: number;
};

export type ListAnalyticsParams = {
platform?: string;
from?: string;
to?: string;
};