# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Social Media Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface Analytics {
  id: number;
  postId: number;
  platform: string;
  likes: number;
  comments: number;
  shares: number;
  reach: number;
  impressions: number;
  clicks?: number;
  recordedAt: Date;
}