# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Social Media Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyticsSummary {
  totalPosts: number;
  totalPublished: number;
  totalScheduled: number;
  totalDrafts: number;
  totalLikes: number;
  totalComments: number;
  totalShares: number;
  totalReach: number;
  totalImpressions: number;
  postsThisWeek?: number;
  engagementRate?: number;
}