# lib/api-zod/src/generated/types/platform.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Social Media Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformName } from './platformName';

export interface Platform {
  name: PlatformName;
  displayName: string;
  connected: boolean;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  connectedAt?: Date | null;
}