# lib/api-zod/src/generated/types/platformConnectInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Social Media Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformConnectInput {
  accessToken: string;
  accountName?: string;
}