# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Social Media Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { PostStatus } from './postStatus';

export interface Post {
  id: number;
  content: string;
  /** @nullable */
  mediaUrl?: string | null;
  platforms: string[];
  status: PostStatus;
  /** @nullable */
  scheduledAt?: Date | null;
  /** @nullable */
  publishedAt?: Date | null;
  createdAt: Date;
  updatedAt?: Date;
  /** @nullable */
  errorMessage?: string | null;
}