# lib/api-zod/src/generated/types/postInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Social Media Automation API
 * OpenAPI spec version: 0.1.0
 */

export interface PostInput {
  /** @minLength 1 */
  content: string;
  mediaUrl?: string;
  /** @minItems 1 */
  platforms: string[];
  scheduledAt?: Date;
}