# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Social Media Automation API
 * OpenAPI spec version: 0.1.0
 */
import type { PostUpdateStatus } from './postUpdateStatus';

export interface PostUpdate {
  /** @minLength 1 */
  content?: string;
  mediaUrl?: string;
  platforms?: string[];
  status?: PostUpdateStatus;
  scheduledAt?: Date;
}